fix: xpath filter passed to 'get_changes'
When an xpath is set then the sysrepo API already provides only filtered changes. Reapplying the same xpath doesn't work when the filter element is not part of the result. Hence, remove all filters when collecting the changes.

For example, the xpath "/ietf-interfaces:interfaces/interface[type='iana-if-type:ethernetCsmacd']/enabled"
subscribes for changes of the 'enabled' leaf but 'type' is not part of the result, in which case get_changes(xpath + '//.', ...) will always be empty!

Hence, remove any filter conditions from the xpath!